feat: Support reconnections

+++ src/EtherSharp/Types/Block.cs
@@ -33,7 +33,7 @@ public record BlockData(
 );
 
 public record BlockDataTrasactionAsString(
-    BigInteger Number,
+    ulong Number,
     DateTimeOffset Timestamp,
     BigInteger? BaseFeePerGas
 );
